backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
psycopg[binary]==3.3.2
python-dotenv==1.2.1
pandas==2.3.3
openpyxl==3.1.5
django-cors-headers==4.9.0
cryptography==46.0.3

